[MFLP-94] feat : 작성자 권한 검증 로직 추가

+++ meongnyang/src/main/java/com/ssafy/meongnyang/api/comment/service/CommentServiceImpl.java
@@ -5,6 +5,8 @@ import com.ssafy.meongnyang.api.comment.dto.request.CommentCreateRequest;
 import com.ssafy.meongnyang.api.comment.dto.request.CommentUpdateRequest;
 import com.ssafy.meongnyang.api.comment.dto.response.CommentGetResponse;
 import com.ssafy.meongnyang.api.comment.repository.CommentRepository;
+import com.ssafy.meongnyang.global.exception.CustomException;
+import com.ssafy.meongnyang.global.response.enums.ErrorCode;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -32,13 +34,20 @@ public class CommentServiceImpl implements CommentService {
         return commentRepository.getCommentById(commentId);
     }
 
-    public void updateComment(CommentUpdateRequest commentUpdateRequest, Long commentId) {
+    public void updateComment(CommentUpdateRequest commentUpdateRequest, Long commentId, Long userId) {
         Comment comment = commentRepository.getCommentById(commentId);
+        if(!comment.getUserId().equals(userId)) {
+            throw new CustomException(ErrorCode.NOT_AUTHOR);
+        }
         comment.setContent(commentUpdateRequest.content());
         commentRepository.updateComment(comment);
     }
 
-    public void deleteComment(Long commentId) {
+    public void deleteComment(Long commentId, Long userId) {
+        Comment comment = commentRepository.getCommentById(commentId);
+        if(!comment.getUserId().equals(userId)) {
+            throw new CustomException(ErrorCode.NOT_AUTHOR);
+        }
         commentRepository.deleteComment(commentId);
     }
 }
